ingestion/requirements.txt
PyYAML==6.0
requests==2.31.0
google-cloud-storage==2.14.0
beautifulsoup4==4.12.2
argparse==1.4.0
pandas==2.0.3
